Improved documentation in the source files.
When someone downloads a source package they should be able to install and
use the software without having to go online. Therefore we must include
installation instructions etc here. The user manual should also be included
here. We need to figure our a way to maintain all documentation in one place
and maybe publish it in many places.