Ensured that event rectangles are never far outside the screen.
This fixes two bugs that both appeared when you had events whose period was
much larger than the current displayed period. The x coordinates for the event
would then be far outside the screen: negative on the left side and large
positive on the right side.

The first bug was that the event rectangles were sometimes not drawn when the
rect passed to the draw method was far outside the screen (-1600 to 35000 for
the x-coordinates for example).

The second bug was that the event balloon was drawn outside the screen. The
balloon was always drawn in the center of the period. But if the center is
outside the screen, the balloon would not be visible. Now when events are never
far outside the screen (maximum MARGIN outside) the center of the period will
always be in the visible area.