Refactor code handling db errors.
New functions in gui utils help enforce the way db errors should be handled and
also remove some incorrect behavior in dialog code (EndModal was sometimes
called even if a dialog had not yet been opened for example).

Dialogs previously displayed a message when a db error occurred and then closed
themselves with an error code. This is changed in this patch so that the main
frame does all the handling of a db error including displaying a message.