Import events into existing categories.
This is how it used to work before. Then it was changed to always create
a completely separate category tree. But that behavior was not desired.
If an imported event has a category that does not exist in the imported
db, that category will be cloned, but it will not keep its parent category.