Improve crash report form.
The body contains indented stacktrace and environment. This
makes it format nicely as markdown when pasted into a ticket.

Notes about steps to reproduce has been removed since it is
mostly never used by user submitting this form. It is better
to ask users for steps to reproduce if you can't figure it
out from the stacktrace.