Problem: Event editor sometimes formats dates incorrectly.
(This is only a problem if the experimental feature 'locale dates' is enabled.)

The date formatter that is used is stored in a global. That makes it hard to
reason about which date formatter we will get when we open the event editor.
It works some times, and others not. (During one test the date formatter
seemed to be incorrectly reset when creating a new timeline.)

Solution: Don't use the global when creating the event editor. Instead get the
date formatter from the config.

(The global is still used in other places which might cause other problems.)