Problem: Some translations do not work because of incorrect shortcuts.
This causes Timeline to crash when started with those translations.

Solution: Manually fix the translation files.

This should really be fixed by removing the shortcuts from the translation
strings. In the meantime, the test added ensures that shortcuts are correct.
Everytime translations are imported, they have to be manually fixed.

Another, temporary solution, is to fix the actual translations on Launchpad.