Problem: Numeric timeline does not use correct time implementation.
At the moment, a number is used. Sometimes it is an int and sometimes
it is a float. Controlling which is hard and would need changes
throughout the codebase.

Solution: Implement a numeric time type (and delta) that can be used
instead of bare numbers.

The solution also involves refactoring the internal time/delta and
also add some missing interface methods.