Change core of data structure (MemoryDB) to be immutable.
This solves two problems:

1. MemoryDB can become corrupt

If for example a circular category was assigned and save_category was never
called, the db would become corrupt. With an immutable core, such a change is
not possible to do without also running the check.

2. Unstable cloning

When a new attribute is added to an event, that attribute also has to be cloned
in order for undo/redo to work. This is difficult to maintain. With an
immutable core, no cloning is needed. This means that undo/redo is trivial to
implement.